Refactor api and api_handlers
Found a good delineator between api and handlers - api is now the filters, handlers is now the reply-ers.  That caused some code move but made the modules more independent.  Also, reduced some verbiage, and genericized with_ in filters to now be add_in.  Modified the CSRF handling so that hopefully the origin/referer check is done before other work, instead of how it was likely done before (after).